requirements.txt
requests==2.31.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0 